RHOAIENG-79877: Extract BreadcrumbItemType, TimeframeTitle, metrics constants to @odh-dashboard/ui-core
