+++ frontend/src/api/prometheus/const.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_LIST_FETCH_STATE } from '#~/utilities/const';
+import { DEFAULT_LIST_FETCH_STATE } from '@odh-dashboard/ui-core/utilities/metrics';
 
 export const PROMETHEUS_BIAS_PATH = '/api/prometheus/bias';
