+++ frontend/src/api/prometheus/serving.ts
@@ -4,8 +4,8 @@ import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
 import { PrometheusQueryRangeResponseDataResult, PrometheusQueryRangeResultValue } from '#~/types';
 import { ModelMetricType } from '#~/pages/modelServing/screens/metrics/ModelServingMetricsContext';
 import { PerformanceMetricType } from '#~/pages/modelServing/screens/types';
-import { RefreshIntervalTitle, TimeframeTitle } from '#~/concepts/metrics/types';
-import { RefreshIntervalValue } from '#~/concepts/metrics/const';
+import { RefreshIntervalTitle, TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
+import { RefreshIntervalValue } from '@odh-dashboard/ui-core/utilities/metrics';
 import useRefreshInterval from '#~/utilities/useRefreshInterval';
 import { PROMETHEUS_BIAS_PATH } from '#~/api/prometheus/const';
 import useQueryRangeResourceData from './useQueryRangeResourceData';
