+++ frontend/src/concepts/distributedWorkloads/DistributedWorkloadsContext.tsx
@@ -4,7 +4,7 @@ import { SupportedArea } from '@odh-dashboard/plugin-core/areas';
 import { getDisplayNameFromK8sResource } from '@odh-dashboard/k8s-core';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
 import { ClusterQueueKind, LocalQueueKind, WorkloadKind } from '#~/k8sTypes';
-import { DEFAULT_LIST_FETCH_STATE } from '#~/utilities/const';
+import { DEFAULT_LIST_FETCH_STATE } from '@odh-dashboard/ui-core/utilities/metrics';
 import { conditionalArea } from '#~/concepts/areas';
 import useSyncPreferredProject from '#~/concepts/projects/useSyncPreferredProject';
 import { ProjectsContext, byName } from '#~/concepts/projects/ProjectsContext';
@@ -15,7 +15,7 @@ import {
   getGenericErrorCode,
   useDWProjectCurrentMetrics,
 } from '#~/api';
-import { RefreshIntervalValue } from '#~/concepts/metrics/const';
+import { RefreshIntervalValue } from '@odh-dashboard/ui-core/utilities/metrics';
 import { MetricsCommonContext } from '#~/concepts/metrics/MetricsCommonContext';
 import PermissionsNotSet from './PermissionsNotSet';
 import useClusterQueues from './useClusterQueues';
