+++ frontend/src/concepts/metrics/kserve/NimMetricsContext.tsx
@@ -12,8 +12,8 @@ import { MetricsCommonContext } from '#~/concepts/metrics/MetricsCommonContext';
 import useKserveMetricsConfigMap from '#~/concepts/metrics/kserve/useKserveMetricsConfigMap';
 import useNimMetricsGraphDefinitions from '#~/concepts/metrics/kserve/useNimMetricsGraphDefinition';
 import useRefreshInterval from '#~/utilities/useRefreshInterval';
-import { RefreshIntervalValue } from '#~/concepts/metrics/const';
-import { RefreshIntervalTitle, TimeframeTitle } from '#~/concepts/metrics/types';
+import { RefreshIntervalValue } from '@odh-dashboard/ui-core/utilities/metrics';
+import { RefreshIntervalTitle, TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
 import { NimMetricGraphDefinition } from '#~/concepts/metrics/kserve/types';
 import { conditionalArea } from '#~/concepts/areas';
