+++ frontend/src/concepts/metrics/kserve/content/KserveMeanLatencyGraph.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { KserveMetricGraphDefinition } from '#~/concepts/metrics/kserve/types';
-import { TimeframeTitle } from '#~/concepts/metrics/types';
+import { TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
 import MetricsChart from '#~/pages/modelServing/screens/metrics/MetricsChart';
 import { useFetchKserveMeanLatencyData } from '#~/api';
 import { convertPrometheusNaNToZero } from '#~/pages/modelServing/screens/metrics/utils';
