+++ frontend/src/concepts/metrics/kserve/content/KserveMemoryUsageGraph.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { KserveMetricGraphDefinition } from '#~/concepts/metrics/kserve/types';
-import { TimeframeTitle } from '#~/concepts/metrics/types';
+import { TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
 
 import { useFetchKserveMemoryUsageData } from '#~/api';
 import MetricsChart from '#~/pages/modelServing/screens/metrics/MetricsChart';
