+++ frontend/src/concepts/metrics/kserve/content/KservePerformanceGraphs.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { KserveMetricGraphDefinition } from '#~/concepts/metrics/kserve/types';
 import { KserveMetricsGraphTypes } from '#~/concepts/metrics/kserve/const';
 import KserveRequestCountGraph from '#~/concepts/metrics/kserve/content/KserveRequestCountGraph';
-import { TimeframeTitle } from '#~/concepts/metrics/types';
+import { TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
 import KserveMeanLatencyGraph from '#~/concepts/metrics/kserve/content/KserveMeanLatencyGraph';
 import KserveCpuUsageGraph from '#~/concepts/metrics/kserve/content/KserveCpuUsageGraph';
 import KserveMemoryUsageGraph from '#~/concepts/metrics/kserve/content/KserveMemoryUsageGraph';
