+++ frontend/src/concepts/metrics/kserve/content/NIMTimePerOutputTokenGraph.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NimMetricGraphDefinition } from '#~/concepts/metrics/kserve/types';
-import { TimeframeTitle } from '#~/concepts/metrics/types';
+import { TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
 import MetricsChart from '#~/pages/modelServing/screens/metrics/MetricsChart';
 import { useFetchNimTimePerOutputTokenData } from '#~/api';
 import { convertPrometheusNaNToZero } from '#~/pages/modelServing/screens/metrics/utils';
