+++ frontend/src/pages/modelServing/customServingRuntimes/CustomServingRuntimeContext.tsx
@@ -4,7 +4,8 @@ import { Bullseye, EmptyState, EmptyStateBody, Spinner } from '@patternfly/react
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import type { TemplateKind } from '@odh-dashboard/k8s-core';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
-import { DEFAULT_LIST_FETCH_STATE, DEFAULT_LIST_WATCH_RESULT } from '#~/utilities/const';
+import { DEFAULT_LIST_FETCH_STATE } from '@odh-dashboard/ui-core/utilities/metrics';
+import { DEFAULT_LIST_WATCH_RESULT } from '#~/utilities/const';
 import { CustomWatchK8sResult } from '#~/types';
 import { useDashboardNamespace } from '#~/redux/selectors';
 import { useTemplates } from '#~/api';
