+++ frontend/src/types.ts
@@ -14,7 +14,6 @@ import type {
 } from '@odh-dashboard/k8s-core';
 import { FeatureFlag } from '@odh-dashboard/plugin-core/areas';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
-import type { EitherNotBoth } from '@odh-dashboard/foundation';
 import { HardwarePodSpecOptions } from '#~/concepts/hardwareProfiles/types';
 import { ImageStreamKind, ImageStreamSpecTagType, NotebookKind } from './k8sTypes';
 
@@ -616,10 +615,6 @@ export type ListWithNonDashboardPresence<T> = {
   items: T[];
   hasNonDashboardItems: boolean;
 };
-
-export type BreadcrumbItemType = {
-  label: string;
-} & EitherNotBoth<{ link: string }, { isActive: boolean }>;
 
 export type DetectedAccelerators = {
   configured: boolean;
