+++ packages/cypress/package.json
@@ -33,6 +33,7 @@
     "@odh-dashboard/model-registry": "*",
     "@odh-dashboard/model-training": "*",
     "@odh-dashboard/plugin-core": "*",
+    "@odh-dashboard/ui-core": "*",
     "@openshift/dynamic-plugin-sdk-utils": "^5.0.1",
     "@testing-library/cypress": "^10.0.1",
     "@testing-library/dom": "^9.3.4",
