+++ packages/model-serving/src/components/metrics/bias/BiasConfigurationPage/BiasConfigurationPage.tsx
@@ -3,7 +3,7 @@ import { Breadcrumb, Button, EmptyStateVariant, PageSection } from '@patternfly/
 import { Link } from 'react-router-dom';
 import { getDisplayNameFromK8sResource } from '@odh-dashboard/k8s-core';
 import { ApplicationsPage } from '@odh-dashboard/ui-core';
-import { BreadcrumbItemType } from '@odh-dashboard/internal/types';
+import { BreadcrumbItemType } from '@odh-dashboard/ui-core/types/metrics';
 import { useModelBiasData } from '@odh-dashboard/internal/concepts/trustyai/context/useModelBiasData';
 import { TrustyInstallState } from '@odh-dashboard/internal/concepts/trustyai/types';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
