+++ packages/model-serving/src/components/metrics/utils.tsx
@@ -4,15 +4,15 @@ import * as _ from 'lodash-es';
 import { BreadcrumbItem } from '@patternfly/react-core';
 import { Link } from 'react-router-dom';
 import { isEnumMember } from '@odh-dashboard/foundation';
-import { TimeframeTitle } from '@odh-dashboard/internal/concepts/metrics/types';
-import { BreadcrumbItemType, PrometheusQueryRangeResultValue } from '@odh-dashboard/internal/types';
+import { TimeframeTitle, BreadcrumbItemType } from '@odh-dashboard/ui-core/types/metrics';
+import { PrometheusQueryRangeResultValue } from '@odh-dashboard/internal/types';
 import {
   BaseMetricRequest,
   BaseMetricRequestInput,
   BiasMetricType,
 } from '@odh-dashboard/internal/api';
 import { BiasMetricConfig } from '@odh-dashboard/internal/concepts/trustyai/types';
-import { PROMETHEUS_REQUEST_RESOLUTION } from '@odh-dashboard/internal/concepts/metrics/const';
+import { PROMETHEUS_REQUEST_RESOLUTION } from '@odh-dashboard/ui-core/utilities/metrics';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import { BIAS_CHART_CONFIGS, BIAS_THRESHOLD_COLOR } from './const';
 import {
