+++ packages/ui-core/package.json
@@ -22,6 +22,7 @@
     "./design/TypedObjectIcon": "./src/design/TypedObjectIcon.tsx",
     "./design/HeaderIcon": "./src/design/HeaderIcon.tsx",
     "./design/TitleWithIcon": "./src/design/TitleWithIcon.tsx",
+    "./utilities/metrics": "./src/utilities/metrics.ts",
     "./utilities/time": "./src/utilities/time.ts",
     "./components/detail": "./src/components/detail/index.ts",
     "./components/detail/OverviewCard": "./src/components/detail/OverviewCard.tsx",
@@ -61,6 +62,7 @@
     "./contexts/NotificationContext": "./src/contexts/NotificationContext.tsx",
     "./contexts/NotificationWatcherContext": "./src/contexts/NotificationWatcherContext.tsx",
     "./types": "./src/types.ts",
+    "./types/metrics": "./src/types/metrics.ts",
     "./components/projectSelector": "./src/components/projectSelector/index.ts",
     "./components/projectSelector/ProjectSelector": "./src/components/projectSelector/ProjectSelector.tsx",
     "./components/projectSelector/ProjectSelectorNavigator": "./src/components/projectSelector/ProjectSelectorNavigator.tsx",
